+++ src/main/java/domain/card/CardDeck.java
@@ -4,9 +4,6 @@ import java.util.ArrayList;
 import java.util.List;
 
 public class CardDeck {
-    private static final int BONUS_THRESHOLD = 11;
-    private static final int ACE_BONUS = 10;
-    private final int INITIAL_HIT_COUNT = 2;
 
     private final List<Card> cards;
 
@@ -16,37 +13,6 @@ public class CardDeck {
 
     public Card hitCard() {
         return cards.removeFirst();
-    }
-
-    public void hitCards(final CardDeck standard) {
-        for (int i = 0; i < INITIAL_HIT_COUNT; i++) {
-            this.addCard(standard.hitCard());
-        }
-    }
-
-    public void addCard(final Card card) {
-        cards.add(card);
-    }
-
-    public int sumWithAce() {
-        int sum = cards.stream()
-                .mapToInt(Card::getScore)
-                .sum();
-
-        if (sum <= BONUS_THRESHOLD && hasA()) {
-            sum += ACE_BONUS;
-        }
-
-        return sum;
-    }
-
-    private boolean hasA() {
-        return cards.stream()
-                .anyMatch(Card::isA);
-    }
-
-    public Card getCardExceptHidden() {
-        return cards.getFirst();
     }
 
     public List<Card> getCards() {
